FIX: Markdown parsing messing up links
The markdown "parser" would get hung up on links followed by a quote due
to the way links were recognized. This commit makes sure links will only
contain allowed characters. In order to do so, the HTML escaping must
not be done before the "parsing" though.

Furthermore these changes seemed to cause too many newlines to be
inserted in the HTML, so as an additional post-processing we now only
replace double-newlines with </br> and remove the single linebreaks.

Fixes #5347