FIX(client): Deadlock on plugin unlink
When a positional data plugin was unlinked while Mumble was playing
audio, a deadlock could be created. This deadlock involved the
m_activePosDataPluginLock and m_positionalData.m_lock. Both locks are
taken in PluginManager::fetchPositionalData. This function calls
PluginManager::selectActivePositionalDataPlugin, but before doing so,
releases the m_activePosDataPluginLock which is re-taken in the
mentioned function.

At the same time the audio thread would call
PluginManager::fetchPositionalData which attempts to take the
aforementioned locks, but this can now lead to a situation in which the
m_activePosDataPluginLock was free and is then taken by the audio thread
but m_positionalData.m_lock is still held by the main thread which is in
the process of calling PluginManager::selectActivePositionalDataPlugin
which then blocks to acquire m_activePosDataPluginLock. This lock is
held by the audio thread but this thread is blocked by waiting until
m_positionalData.m_lock is available again. Thus, a deadlock is created.

This commit fixes this situation by making sure m_positionalData.m_lock
is released before calling selectActivePositionalDataPlugin.

Fixes #5319