REFAC(server): Improve semantics in msgUserState
Inside Server::msgUserState there are a couple of places that implement
the functionality of a client making modifications to itself (e.g.
self-mute). In these cases, uSource == pDstServerUser. Therefore
functionality-wise, it doesn't matter which of these variables the
implementation uses. However, there is a semantic differences and as
such this commit refactors these code parts to perform the actions on
pDstServerUser instead of uSource in order to make clear that
pDstServerUser is always the user to whom the action is applied.