FIX(client): very distorted RNNoise output
When the input signal is very loud and clipping, the output of RNNoise
is very distorted, way more than the input signal.

This is because RNNoise outputs a signal that is unbounded and sometimes
goes over the limits of the signed 16-bit signal that Mumble uses
internally. When Mumble converts RNNoise output signal to 16-bit,
integer overflow occurs and creates a very loud distortion.

Mumble now clamps the output of RNNoise which results in a signal that
is not more distorted than the input signal.

Fixes #4391