CHANGE(client): Remove channel listener persistence
This is done, since the underlying framework for how channel listeners
are handled (more specifically: where they are stored) will be changed
in an upcoming PR. This upcoming change will cause listener persistence
to only apply to registered users. Thus, if the current persistence
feature was to be retained, that would mean that when introducing this
change, we'd remove a feature from Mumble.

As that is not a great thing to do, we'll remove the feature now that it
is still unreleased and then the upcoming change will _introduce_
persistence for some users instead of taking it away from others.