MAINT: Add script to sign macOS binaries and packages
Right now we are using "osxdist.py" to generate a DMG package from scratch.
The script also takes care of signing the content.

However, our "mumble-releng" repository has had a dedicated script called "sign-dmg.py" since 2013.
In fact, that's what we have always been using to sign the app bundle.

As we're planning to eventually ditch "osxdist.py" in favor of CPack's generator, we decided to complete "sign-dmg.py".
The script now:

- Takes care of signing all binaries in the bundle in addition to it, effectively replacing "osxdist.py".
- Supports single binaries, replacing "sign-mach-o.py" (which only called codesign() from "sign-dmg.py").
- Supports PKG packages.
- Provides a "--config" option to specify the path to the configuration file. Default: "$HOME/.sign_macOS.cfg".
- Provides a "--entitlements" option to specify the path to the plist file containing the requested capabilities.
  This is mandatory for notarization to succeed. More specifically, the "--options runtime" codesign parameter is.
  The parameter is not passed if this option is not used because it would cause Mumble to crash upon audio input use.

The name is changed to "sign_macOS.py" as the script does not only handle DMG packages anymore.