REFAC(positional-audio): Remove useless memory locations in GTAV
These addresses provide no additional meaningful data. The identity
context is supposed to identify an user, not provide meaningless info
about unrelated ingame objects.

Remove these unnecessary variables, as they only increase the amount of
work needed when reverse-engineering positional audio.

The username is already alone an unique identity, so instead of setting
the identity to a JSON of multiple values, we now set the identity to
just the username.