CI(appveyor): Enable shortcut elevation for signed builds
We can't enable the elevation option by default, since it can only be
used when the binary is properly signed (otherwise Windows would just
hard-crash the application).

However, we're only using AppVeyor to build release artifacts and have
the signing process included in the automation. Therefore, it makes
sense to enable this option on all AppVeyor builds.