FIX(client): Context and identity not getting cleared
If a plugin loses link, the context and identity that have been
broadcast to the server are not cleared.

This commit makes sure that an additional message is sent to the server
that does take care of the clearing.

Fixes #5466

(cherry picked from commit a7e8a20af2da06ebde1eb59c3fac12b3c37b543e)