MAINT(manpages): Replace `murmur` with `mumble-server`
The name `murmur` is deprecated and shouldn't be used anymore, as
decided by the Mumble project. While the server binary already uses the
"new" name, the man pages don't. Thus, rework the man pages so that its
successor `mumble-server` instead `murmur` is used and rename the
filenames of the man pages accordingly.

Also, rename the script `murmur-user-wrapper` to
`mumble-server-user-wrapper`, since also its man page is renamed.

This does not affect the configuration file `murmur.ini`.

Signed-off-by: Felix Singer <felixsinger@posteo.net>