FIX(client,ui): PFS information being incorrect
The information about perfect forwards secrecy (PFS) displayed within
the server information dialog (TCP tab) was missing for newer ciphers.
The reason was, that we are using a hard-coded list of cipher
information and the new ones (TLS 1.3) are not part of that.

However, in the meantime Qt has implemented indirect support for
checking whether PFS is used, so we change the implementation to make
use of that instead.

Fixes #5469

(cherry picked from commit 379140fb56df2c102dde84a174138574933ca58f)