CHANGE(client): Don't use hard-coded cipher info
PR #1825 introduced a hard-coded list of cipher information in order to
be able to display extended information about the encryption details of
a connection. Keeping this information up-to-date had to be done by
means of a dedicated python script.

The problem with this approach is that
a) hard-coding something like this is never a good idea
b) the updating has been forgotten in the last several years, leading to
outdated information
c) the update script is currently not working anymore

Additionally, the current code only used the extended information to
query the perfect forward secrecy support anyway and that has been
implemented in a different way in
de7b02fb3c3c8756377fd789006465bf2bd86ffd.

The only other thing we used the extended info for, was to obtain the
RFC name of ciphers, but since the list was outdated anyway, we nowadays
probably used the fallback (using the name provided by Qt) more often
than not.

Therefore the hard-coded information has been removed and we now simply
always stick to the cipher name as reported by Qt. If someone
desperately requires the RFC names, they'll have to take it up with Qt
to expose this name in their API as curating a hard-coded list in our
code base is not a viable option.