FIX(client): DB not found after migration on Windows
When the DB is migrated, it won't be found anymore after a restart,
because the new DB location was not saved in the settings.

This commit makes sure that during the migration, the path stored in the
settings gets updated and then subsequently saved.

Because of this, the dedicated migration code block in the Database
class is no longer needed (at the point it runs, the DB setting has
already been migrated) and was therefore removed.