FIX(client): Plugin settings not getting migrated
When migrating the data directory, the plugin settings have been let
untouched. However, since these contain the plugin's absolute path (both
as a immediate property and also in the hash identifying the plugin),
the settings wouldn't apply to the migrated plugins. Therefore, the
migration would lead to a loss of plugin settings.

This commit adds dedicated code that remedies this situation by
explicitly adapting the plugin settings to the migrated plugin paths.