BUILD(cmake): Build/install celt as module
It loaded at runtime using dlopen(3) thus should be of type `MODULE`
not `SHARED` (dynamically linked) as per cmake docs[0].

This makes cmake's `install()` create an unversioned file as expected,
e.g. libcelt0.so on OpenBSD.

Account for all existing module names without versions now.

0: https://cmake.org/cmake/help/latest/command/add_library.html#id2

PP: This came up on OpenBSD where shared library versions are controlled
by the ports infrastructure and not the software itself;  improving
automated cmake handling showed that a) the file libcelt0.so.0.7.0 had
a wrong version scheme (OpenBSD uses .so.<maj>.<min>) and b) is not a
dynammically linked library but just a module.

Thus building celt as a module and shipping libcelt0.so aligns with
OpenBSD's ports infrastructure and the way mumble actually uses celt.