FIX(client): Audio wizard being shown repeatedly
The switch to a new settings format messed up the setting for whether or
not the audio wizard has been shown already or not. For new
installations, the wizard would be shown, but the fact that it has been
shown is not reflected in the settings causing the wizard to be re-shown
upon a restart.

This commit fixes this by making sure that the respective setting is set
after the wizard has been shown for the first time.