FIX(client): resolve list tags, etc. polluting log
As reported in #4491, #4986 and #5430, since the changes to the chat in
the 1.4.x release, some tags would "pollute" the rest of the log, i.e.
cause future log entries to be contained within them.

The source of the issue seems to be that the `insertBlock` method of
`QTextCursor` appears to correspond to the `<p>` HTML tag. The `<p>` tag
may only contain inline elements, but `QTextEdit` will not fail outright
when inserting block-level elements and will instead attempt to correct
the invalid input which results in the behaviour reported in the above
issues (as far as I can tell).

My proposed solution is to use the `insertFrame` method for all
messages.

Fixes #4491

(cherry picked from commit ae097d6354d99325093dd5a12cb8632e2ba20edf)