FIX(client): Prevent endless loop when disconnecting PipeWire stream
Since version 0.3.51, PipeWire has support for buffers with a size of 0
being passed in the process callback. More specifically, this feature
was added in commit 96286fb8b11658a0fdaa61194504a3f9541b25e6 ("resample:
use a -1 buffer size to drain") in PipeWire [1].

Unfortunately, this causes an endless loop, because we try to destroy
the PipeWire stream, but it keeps waiting for the node to be completely
drained, which now never happens because it keeps calling
processCallback() where we feed buffers with a size of 0 to it.

To fix this problem, we do not longer pass 0-size buffers to PipeWire.
Instead we now push silence, if we don't have any data available. This
is what PipeWire prefers anyway, the process callback wants a buffer to
be filled. If we don't do that, the node will stay in
SPA_STATUS_NEED_DATA state and even if we ignore the obvious endless
loop error, this can cause other problems.
For example above mentioned commit was fixed in PipeWire with
32e957345d63a6d6e3d223057d764ff661f5d101 ("audioadapter: don't loop
forever") [2], which is better than the endless loop but still
preventable.
So in order to avoid unnecessary xruns, we push silence instead of empty
buffers.

Fixes #5647

[1] https://gitlab.freedesktop.org/pipewire/pipewire/-/commit/96286fb8b11658a0fdaa61194504a3f9541b25e7
[2] https://gitlab.freedesktop.org/pipewire/pipewire/-/commit/32e957345d63a6d6e3d223057d764ff661f5d101