BUILD(macos): Fix plugins not getting bundled
If the osxdist script was not executed within the build directory, it
would not bundle the built plugins (most notably the Link plugin) as it
was using a relative path and was ignoring the binary dir command-line
option.

This is remedied by this commit.

Fixes #5571

(cherry picked from commit 4e2e53d0f92a0f92d724ebbd15689e682fa04c83)