CHANGE(client): Don't hard block shortcuts on Wayland
In #5307 we disabled global shortcuts on Wayland, but we did it in a way
that would prevent anyone from re-enabling the support.

Since we have received reports of some users actually making use of the
shortcut system on Wayland, even without proper Wayland support, we
decided to remove the hard-lock and allow these users to re-enable the
shortcuts, if they so want.

Fixes #5454