FEAT(client, ui): separate ipv4 and ipv6 in ServerItem tooltip
The "Addresses" field in the ServerItem tooltip shows both ipv4 and ipv6
addresses. The order of addresses in this field is not determined.
In the wild it is randomized. As discussed in #5696 it is better to implement
a concrete ordering to prevent user confusion. To further improve the
readability of the addresses we split the addresses field in two fields
called "IPv4 address" and "IPv6 address". The ipv4 address is now always
shown above the ipv6 address.

Implements #5696

Co-Authored-By: Thomas Pawelek <thomas.pawelek@stud.uni-hannover.de>