BUILD(cmake, overlay): Include missing CMake module
We use the CHECK_INCLUDE_FILE macro without previously including the
corresponding CMake module. This usually still works as the Opus CMake
files include the same module and apparently are loaded before
processing the overlay CMakeLists.txt. Thus, thee Opus code fixes this
issue for us by accident.

However, when building without the bundled Opus version, this leads to a
build error as in this case the Opus CMake scripts are not processed by
us.

Fixes #5742