FIX(client,ui): Show friendlier text when displaying notifications
OS-level notifications that are shown when a message is received use a plain-text representation of the sent message. If a message is sent that cannot be gracefully converted to plain text, such as from a pasted image or other embedded content, the Unicode Object Replacement Character is shown instead.

To make the notification body more user-friendly, a more suitable representation of that embedded content is now displayed.

Fixes mumble-voip#5764