FIX(client, ui): Don't display unsupported ACL
If a recent client (1.4 or newer) was to connect to a pre-1.4 server and
decide to edit a channel's ACL, the user would be presented with the
option to configure the Listen ACL. However, since the server is too old
to know about this particular ACL, setting it will have no effect. Thus,
to the end user this creates the impression that something is broken.

This commit ensures that the Listen ACL is only shown, if the server one
is currently connected to, is recent enough to know about this ACL.

Fixes #5692

(cherry picked from commit d1f5be93d5dc65d09dfb253e7ba251b2a9f1f99c)