FIX(server): Update suppress state on ACL change
The suppress state of a user represents whether the user is currently in
a channel in which they don't have the Speak ACL. As of now this state
is only updated whenever the user switches channel.

However, the suppress state should also change, if the ACL for a given
channel change. This is what the current commit does. If this is not
done, a user will be able to continue speaking in a channel where they
should not be able to and only if they leave and re-join the channel
will the new ACL take effect.

To avoid such a situation, the current code updates the suppress state
every time the ACL change.

Fixes #5264

(cherry picked from commit 83d1e311c9f37d93f999e667b505d2df7e334018)