CHANGE(ui): Show action icons in menu bar where applicable
Since the introduction of the menu bar, action icons were hidden in it.

This commit enables the display of existing icons for actions which are not
checkable (otherwise it would clash with the check-mark). Additionally, the
"heart" icon is enabled for "Add Friend..." and the "i" icon is used for user
"Information..."