FIX(ui): User and channel status icons take display scaling into account
Previously, since the inception of the in-line status icons (b39f73befca)
the status icon size has been hard-coded. In contrast to pretty much all other
icons this had the effect, that user display scaling was not taken into account and
the status icons appeared too small, if the scaling was >1.0x

This commit reverts the static constants for the icon size introduced in 42d74df5ec
and calculates the values based on the font size. (This is also done in the
TalkingUI for example). In doing so, the icons will resize with the font and therefore with
display scaling.

Fixes #5632