FEAT(client, ui): Show minimal view note while unconnected
When not connected to a server, the minimal view will not display any UI
elements, which can be confusing (especially since the minimal view
state is saved and restored on restart).

In order to let users know what is going on and how to make the UI
elements re-appear, this commit adds a new note to the UI in these
cases.

Fixes #5495