FIX(ui): Use correct font size for scaling status icons
Previously (#5772), we implemented a fix to set the status icon size
according to the user display scaling. The calculation is based upon
font sizes. However, due to an oversight, the wrong font size was used
as the base for this calculation. However, the problem was not visible,
if by coincidence the correct font size was the same as the one used by
accident.

This commit changes the calculation of the icon size once again to use
the correct font information.

Fixes one part of #5817

(cherry picked from commit a587514c2f49dbc65b8cb7c9710fe183b2c750a6)