FIX(client, ui): Make hiding UI elements persistent again
Due to a regression introduced in
6aa88691e16c807b998301f2cbecfda599ef4592 when hiding individual UI
components (e.g. the icon toolbar) - possible when using a custom
layout - these elements would re-appear after restarting Mumble (or
after applying the settings or after exiting minimal view).

The reason is that the state of whether or not a given UI component is
shown or not is saved as the window's geometry. In the above commit we
added code that explicitly shows or hides these elements after the
geometry has been restored, thus overwriting the visibility changes of
this geometry restoration.

The fix for this problem is to only explicitly hide the elements when
entering minimal view but let the showing be handled by restoring the
window's geometry. Thus, only those elements are shown again, that have
previously been visible as well (and thus have not been explicitly
hidden by the user).

Fixes #5832