FIX(client, ui): Don't allow manual toggling of minimal view note
In 39047d2a57a9288262c5f303d96798a1b1692b8d a new note was added for
when a user enters minimal view but is not connected to a server.
Incidentally, this caused the new note to appear in the popup menu
appearing as a context menu when using a custom layout, which in theory
would allow a manual toggling of that note's visibility. However, this
is not what we want and thus this commit makes sure that this popup menu
only contains entries for UI elements that we really want to be manually
toggelable.

(cherry picked from commit b3a9fba786020489954176bfacec75ae2c9f59ef)