FIX(server): Look for mumble-server* files
Previously the server was only looking for files named e.g. murmur.ini
or murmur.sqlite. With this commit, the server will now look first for
files called mumble_server.ini and mumble_server.sqlite before falling
back to the old names (for backwards compatibility).

This is meant as a means to drive forward the server renaming from
murmur(d) to mumble-server.