BUILD(cmake): Always keep absolute install paths around
Previously we would define our install paths as relative paths when
using the packaging option and turning them into absolute paths
otherwise.

While relative paths are required for CPack to work properly, we use
the paths e.g. for passing them in via macros and there we absolutely
require the full, absolute paths to be passed.

Thus, this commit removed the discrepancy of the results when built with
packagin=ON vs packagin=OFF and now always defines the install paths as
relative paths. However, we now keep an additional set of paths around
that are now always absolute.