FIX(client): Explicitly store shortcut type in DB
When storing shortcuts in the database (which happens only for
server-specific shortcuts), the assumption has been that the type of the
stored shortcut is always "Whisper/Shout" because up to now this is the
only shortcut that can be server-specific.

However, this approach is not very future-proof, which is why this
commit makes sure that from now on, we explicitly store the type of the
shortcut as well.