FIX(client): Clear shortcut data also for user-types
Up to now the logic for clearing shortcut data upon a change in the
shortcut config window was to see if the type() of the current data
matches that of the shortcut's default data. However, for user-defined
types QVariant::type() always returns QVariant::UserType and thus
different user-defined types could not be distinguished.

If we instead use QVariant::userType(), this problem does no longer
exist.