FIX(client): Fixes link context menu creation for channels
The commit 902fe5841ca03f59b introduced links in the log for users
and channels. This either never worked correctly, or regressed due
to an undocumented behavior change in Qt. The 'channelid' stored in
the host part of a QUrl is automatically converted to an IP address
(e.g. '1' -> '0.0.0.1') which lets the channel lookup fail.
A different code path for clients was present, so the issue was not
observed often (or at all) there.

This commit changes the host part composition to contain a string
(e.g. '1' -> 'id.1') and makes sure this new string is parsed correctly.

Fixes #2961