FIX(server): Explicitly set parsed timestamps to UTC for rememberchannelduration
In ccbf4072c the rememberchannelduration logic was improved. An explicit
call to 'setTimeSpec(Qt::UTC)' was removed, because it was assumed that
'secsTo' would already be sufficient as it converts both inputs to UTC
automatically. Due to an oversight this lead to a regression.

The reason setTimeSpec is acutally needed is because the DB already
contains UTC values. The setTimeSpec call tells Qt that the loaded
values are actually already UTC and do not have to be converted.
Without the call to setTimeSpec, the DB values are effectively
converted twice, which gives rememberchannelduration an offset
based on the server local time.